5.4.0 engine changes
